Extract shared assertSourceIsReferenceable helper
This assertion was being repeated in a few places, and so it felt like
the right time to DRY that up.
